// apps/backend/src/common/guards/jwt-auth.guard.d.ts
import { ExecutionContext } from '@nestjs/common';
import { Reflector } from '@nestjs/core';
import { Observable } from 'rxjs';
declare const JwtAuthGuard_base: import("@nestjs/passport").Type<import("@nestjs/passport").IAuthGuard>;
export declare class JwtAuthGuard extends JwtAuthGuard_base {
    private reflector;
    constructor(reflector: Reflector);
    canActivate(context: ExecutionContext): boolean | Promise<boolean> | Observable<boolean>;
    handleRequest<TUser = any>(err: unknown, user: TUser, info: unknown, context: ExecutionContext, status?: unknown): TUser;
}
export {};
//# sourceMappingURL=jwt-auth.guard.d.ts.map