// apps/backend/src/products/dto/filter-products.dto.d.ts
export declare class FilterProductsDto {
    status?: string;
    brandId?: string;
    categoryId?: string;
    catalogId?: string;
    productType?: string;
    search?: string;
    audience?: string;
    formFactor?: string;
    signalProcessing?: string;
    powerLevel?: string;
    hearingLossLevel?: string;
    smartphoneCompatibility?: string;
    paymentOption?: string;
    availabilityStatus?: string;
    limit?: number;
    offset?: number;
    sort?: 'newest' | 'price_asc' | 'price_desc';
}
//# sourceMappingURL=filter-products.dto.d.ts.map